# docs/swagger.yaml
definitions:
...
  requests.CreateInterestsTemplate:
    properties:
      interests:
        items:
          type: string
        type: array
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - interests
    - title
    type: object
...
  requests.DeleteInterestsTemplate:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  requests.EditInterestsTemplate:
    properties:
      id:
        type: integer
      interests:
        items:
          type: string
        type: array
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - id
    - interests
    - title
    type: object
...
  responses.CreateInterestsTemplateDTO:
    properties:
      status:
        type: string
      task_template:
        $ref: '#/definitions/responses.InterestsTemplateDTO'
    type: object
...
  responses.DeleteInterestsTemplateDTO:
    properties:
      status:
        type: string
    type: object
...
  responses.EditInterestsTemplateDTO:
    properties:
      status:
        type: string
      task_template:
        $ref: '#/definitions/responses.InterestsTemplateDTO'
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  responses.GetAllInterestsTemplatesDTO:
    properties:
      task_templates:
        items:
          $ref: '#/definitions/responses.InterestsTemplateDTO'
        type: array
    type: object
...
  responses.GetInterestsTemplateDTO:
    properties:
      task_template:
        $ref: '#/definitions/responses.InterestsTemplateDTO'
    type: object
...
  responses.InterestsTemplateDTO:
    properties:
      id:
        type: integer
      interests:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  responses.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/template/interests/all:
    get:
      consumes:
      - application/json
      description: Retrieves all interest templates for a specific user, with pagination
        support.
      parameters:
      - description: The page offset for pagination. Default is 0.
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved interest templates
          schema:
            $ref: '#/definitions/responses.GetAllInterestsTemplatesDTO'
        "400":
          description: Invalid token or offset
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: No task templates found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get all interest templates by user ID
      tags:
      - InterestsTemplates
  /api/template/interests/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a specific interests template, provided the user is the
        author.
      parameters:
      - description: Template ID
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.DeleteInterestsTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted template
          schema:
            $ref: '#/definitions/responses.DeleteInterestsTemplateDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Access forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete Interests Template
      tags:
      - Interests Template
  /api/template/interests/edit:
    put:
      consumes:
      - application/json
      description: Updates the title or list of interests of an existing template.
      parameters:
      - description: Updated template data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.EditInterestsTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated template
          schema:
            $ref: '#/definitions/responses.EditInterestsTemplateDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Access forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Edit Interests Template
      tags:
      - Interests Template
  /api/template/interests/get/{id}:
    get:
      description: Fetches a specific interests template by its ID, provided the user
        is the author.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved template
          schema:
            $ref: '#/definitions/responses.GetInterestsTemplateDTO'
        "400":
          description: Invalid token or template ID
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Access forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Get Interests Template
      tags:
      - Interests Template
  /api/template/interests/new:
    post:
      consumes:
      - application/json
      description: Creates a new template based on the provided title and list of
        interests.
      parameters:
      - description: Template data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateInterestsTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created template
          schema:
            $ref: '#/definitions/responses.CreateInterestsTemplateDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create Interests Template
      tags:
      - Interests Template
swagger: "2.0"